+++ egui/src/containers/collapsing_header.rs
@@ -24,7 +24,7 @@ impl Default for State {
 
 impl State {
     pub fn from_memory_with_default_open(ctx: &Context, id: Id, default_open: bool) -> Self {
-        *ctx.memory().data.get_or_insert_with(id, || State {
+        *ctx.memory().id_data.get_or_insert_with(id, || State {
             open: default_open,
             ..Default::default()
         })
@@ -35,7 +35,10 @@ impl State {
         if ctx.memory().everything_is_visible() {
             Some(true)
         } else {
-            ctx.memory().data.get::<State>(id).map(|state| state.open)
+            ctx.memory()
+                .id_data
+                .get::<State>(id)
+                .map(|state| state.open)
         }
     }
 
@@ -292,7 +295,7 @@ impl CollapsingHeader {
                 })
                 .inner
             });
-            ui.memory().data.insert(id, state);
+            ui.memory().id_data.insert(id, state);
 
             if let Some(ret_response) = ret_response {
                 CollapsingResponse {
