+++ egui/src/grid.rs
@@ -59,7 +59,7 @@ pub(crate) struct GridLayout {
 
 impl GridLayout {
     pub(crate) fn new(ui: &Ui, id: Id) -> Self {
-        let prev_state = ui.memory().data.get_or_default::<State>(id).clone();
+        let prev_state = ui.memory().id_data.get_or_default::<State>(id).clone();
 
         // TODO: respect current layout
 
@@ -212,7 +212,7 @@ impl GridLayout {
         if self.curr_state != self.prev_state {
             self.ctx
                 .memory()
-                .data
+                .id_data
                 .insert(self.id, self.curr_state.clone());
             self.ctx.request_repaint();
         }
